Extract helpers into separate modules for ease of testing
